requirements.txt
asgiref==3.3.1
Django==3.1.7
djangorestframework==3.12.2
gunicorn==20.0.4
imutils==0.5.4
numpy==1.20.1
opencv-python==4.5.1.48
Pillow==8.1.2
pytz==2021.1
scipy==1.6.1
sqlparse==0.4.1
